Moved widget creation and utility functions to a new `import_widgets` module
